feat(style): make the result panel way more fluid
