+++ src/components/organisms/AnalyzeStep/AnalyzeStep.module.css
@@ -25,13 +25,26 @@ div.rowGood {
   list-style: none;
   padding: 0;
 }
-.emptyPlaceholder {
-  display: flex;
+.tableWrap {
+  position: relative;
+}
+.emptyFallback {
   font-size: 1rem;
-  font-style: italic;
-  opacity: 0.3;
+  opacity: 0.5;
   color: var(--color-text-secondary);
   padding: var(--space-xl) 0;
+}
+.loadingFallback {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: 10;
+  justify-content: center;
+  align-items: center;
+  display: flex;
+  backdrop-filter: blur(2px);
 }
 
 @media (max-width: 768px) {
