+++ src/styles/globals.css
@@ -3,7 +3,7 @@
   --on-primary: #eee;
   --secondary: #eee;
   --on-secondary: #222;
-  --error: #f00;
+  --error: #d42b2b;
   --on-error: #eee;
   --success: #1ade1a;
   --on-success: #222;
